pyproject.toml
[build-system]
requires = [
    "setuptools>=45.0",
    "setuptools_scm[toml]>=6.2",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "austaltools"
dynamic = ["version"]
description = "Tools for use with the Lagrangian atmospheric pollutant dispersion model AUSTAL (AUSbreitungsmodell nach TA-Luft)."
authors = [
    {name = "Clemens Drüe", email = "druee@uni-trier.de"},
]
license = {text = "EUPL-1.2"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["meteorology", "air quality", "modeling"]
dependencies = [
    "cdsapi",
    "gdal",
    "matplotlib",
    "meteolib",
    "netcdf4",
    "numpy",
    "pandas",
    "pyyaml",
    "readmet",
    "requests",
    "setuptools",
    "tqdm",
    "urllib3",
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-argparse",
]
test = [
    "pytest",
    "scipy",
]
dev = [
    "flake8",
    "pytest",
    "scipy",
    "tox",
]

[project.urls]
Homepage = "https://github.com/cdruee/austaltools"
Repository = "https://github.com/cdruee/austaltools"
Issues = "https://github.com/cdruee/austaltools/issues"
Documentation = "https://austaltools.readthedocs.io"

[project.scripts]
austaltools = "austaltools.command_line:main"
austal-input = "austaltools.austal_input:main"
configure-austaltools = "austaltools.configure_austaltools:main"

[tool.setuptools]
packages = ["austaltools", "austaltools.data"]
include-package-data = true
zip-safe = false

[tool.setuptools.package-data]
"austaltools.data" = ["*.txt", "*.json"]

[tool.setuptools_scm]
write_to = "austaltools/_version.py"
write_to_template = '''"""Version information generated by setuptools_scm"""

__version__ = "{version}"
'''
fallback_version = "0.0.0.dev0"
local_scheme = "no-local-version"

[tool.setuptools.cmdclass]

[tool.flake8]
extend-exclude = ["pub", "docs", "build", "deb_dist", "venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.tox]
env_list = ["py", "flake8", "docs"]

[tool.tox.env_run_base]
deps = [
    "pytest",
    "numpy",
    "pandas",
    "readmet",
    "scipy",
    "gdal",
    "matplotlib"
]
commands = ["pytest {posargs}"]

[tool.tox.env.flake8]
deps = ["flake8"]
commands = ["flake8 {posargs}"]

[tool.tox.env.docs]
deps = [
    "sphinx",
    "sphinx-argparse"
]
commands = [
    "sphinx-build -W -b html docs docs/_build/html"
]